add dateFormatPipe
Make sure to use a pipe for static functionality instead of using a function, since this will be triggered for every change detection cycle
